+++ Pandora/Paradigm/Basis.hs
@@ -21,8 +21,8 @@ import Pandora.Paradigm.Basis.Product as Exports
 import Pandora.Paradigm.Basis.Constant as Exports
 import Pandora.Paradigm.Basis.Identity as Exports
 
-import Pandora.Core.Functor (type (~>))
 import Pandora.Core.Morphism ((!))
+import Pandora.Core.Transformation (type (~>))
 
 note :: e -> Maybe ~> Conclusion e
 note x = maybe (Failure x) Success